Backend/requirements.txt
flask
flask-smorest
python-dotenv
sqlalchemy
flask-sqlalchemy
email_validator
passlib
flask-jwt-extended
pymysql
apscheduler
flask-restful
Flask-Mail
celery
redis
